+++ include/transform.hpp
@@ -31,7 +31,7 @@ public:
     {
     }
 
-    virtual bool intersect(const Ray &r, Hit &h, float t_min)
+    virtual bool intersect(const Ray &r, Hit &h, double t_min)
     {
         Vector3f trSource = transformPoint(transform, r.getOrigin());
         Vector3f trDirection = transformDirection(transform, r.getDirection());
